requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asttokens==3.0.0
asyncpg==0.30.0
attrs==25.3.0
bcrypt==4.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
dataclasses-json==0.6.7
decorator==5.2.1
Deprecated==1.2.18
executing==2.2.0
fastapi==0.115.12
fastapi-pagination==0.12.34
frozenlist==1.6.0
googleapis-common-protos==1.70.0
greenlet==3.2.0
grpcio==1.71.0
grpcio-tools==1.71.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
ipython==9.1.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jsonpatch==1.33
jsonpath-ng==1.7.0
jsonpointer==3.0.0
langchain==0.3.23
langchain-community==0.3.21
langchain-core==0.3.54
langchain-text-splitters==0.3.8
langgraph==0.3.31
langgraph-checkpoint==2.0.24
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.61
langsmith==0.3.32
Mako==1.3.10
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
multidict==6.4.3
mypy-extensions==1.0.0
numpy==2.2.4
orjson==3.10.16
ormsgpack==1.9.1
packaging==24.2
parso==0.8.4
pexpect==4.9.0
ply==3.11
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==5.29.4
psycopg2==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.10.1
python-dotenv==1.0.1
python-yandex-cloud-logging==0.1.0
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
setuptools==78.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
stack-data==0.6.3
starlette==0.46.2
tenacity==9.1.2
traitlets==5.14.3
typing-inspect==0.9.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.0
wcwidth==0.2.13
wrapt==1.17.2
xxhash==3.5.0
yaml-settings-pydantic==2.3.2
yandexcloud==0.340.0
yarl==1.20.0
zstandard==0.23.0
